Move Tag tests to integrations package
